Adding induvidual chart controls, adding expand chart modal, fixing disabled formcontrol text color.
