+++ SpeedRunApp.UI/src/components/SpeedRunGridChartVue.vue
@@ -10,46 +10,58 @@
                 </div>
             </h5>
         </div>
-        <div class="container row chart-container" :style="[ showcharts ? null : { display:'none' } ]">
-            <div class="col-sm-4" style="min-height:300px;">
-                <div v-if="loading" class="d-flex" style="height:100%;">
-                    <div class="m-auto">
-                        <i class="fas fa-spinner fa-spin fa-lg"></i>
+        <div class="container" :style="[ showcharts ? null : { display:'none' } ]">
+            <div class="row">
+                <div class="col-lg-4" style="min-height:300px;">
+                    <div v-if="loading" class="d-flex" style="height:100%;">
+                        <div class="m-auto">
+                            <i class="fas fa-spinner fa-spin fa-lg"></i>
+                        </div>
+                    </div>
+                    <div v-else @click="onChartClick($event, 1)" class="expandable" style="height:100%;">
+                        <worldrecord-chart chartconainerid="divChart1" :tabledata="tabledata" :isgame="isgame" :title="title" :istimerasc="istimerasc"></worldrecord-chart>                
                     </div>
                 </div>
-                <div id="divChart1" :style="[ loading ? { display:'none' } : null ]"></div>
-            </div>
-            <div class="col-sm-4" style="min-height:300px;">
-                <div v-if="loading" class="d-flex" style="height:100%;">
-                    <div class="m-auto">
-                        <i class="fas fa-spinner fa-spin fa-lg"></i>
+                <div class="col-lg-4" style="min-height:300px;">
+                    <div v-if="loading" class="d-flex" style="height:100%;">
+                        <div class="m-auto">
+                            <i class="fas fa-spinner fa-spin fa-lg"></i>
+                        </div>
+                    </div>
+                    <div v-else @click="onChartClick($event, 2)" class="expandable" style="height:100%;">
+                        <speedrun-percentile-chart chartconainerid="divChart2" :tabledata="tabledata" :isgame="isgame" :title="title" :istimerasc="istimerasc"></speedrun-percentile-chart>                
                     </div>
                 </div>
-                <div id="divChart2" :style="[ loading ? { display:'none' } : null ]"></div>
-            </div>
-            <div class="col-sm-4" style="min-height:300px;">
-                <div v-if="loading" class="d-flex" style="height:100%;">
-                    <div class="m-auto">
-                        <i class="fas fa-spinner fa-spin fa-lg"></i>
+                <div class="col-lg-4" style="min-height:300px;">
+                    <div v-if="loading" class="d-flex" style="height:100%;">
+                        <div class="m-auto">
+                            <i class="fas fa-spinner fa-spin fa-lg"></i>
+                        </div>
                     </div>
-                </div>              
-                <div id="divChart3" :style="[ loading ? { display:'none' } : null ]"></div>
-            </div>
+                    <div v-else @click="onChartClick($event, 3)" class="expandable" style="height:100%;">
+                        <top-speedrun-chart chartconainerid="divChart3" :tabledata="tabledata" :isgame="isgame" :title="title" :istimerasc="istimerasc"></top-speedrun-chart>                
+                    </div>
+                </div>
+            </div>            
         </div>
+        <modal v-if="showChartModal && !loading" contentclass="cmv-modal-lg" bodyclass="p-0" @close="showChartModal = false">
+            <template v-slot:title>
+                {{ chartModalTitle }}
+            </template>
+            <div v-if="selectedChartID == 1">            
+                <worldrecord-chart chartconainerid="divChartModal" :tabledata="tabledata" :isgame="isgame" :title="title" :istimerasc="istimerasc" :ismodal="true"></worldrecord-chart>                
+            </div>
+            <div v-else-if="selectedChartID == 2">
+                <speedrun-percentile-chart chartconainerid="divChartModal" :tabledata="tabledata" :isgame="isgame" :title="title" :istimerasc="istimerasc" :ismodal="true"></speedrun-percentile-chart>                
+            </div>
+            <div v-else-if="selectedChartID == 3">
+                <top-speedrun-chart chartconainerid="divChartModal" :tabledata="tabledata" :isgame="isgame" :title="title" :istimerasc="istimerasc" :ismodal="true"></top-speedrun-chart>                
+            </div>            
+        </modal>              
     </div>
 </template>
 <script>
     import axios from 'axios';
-    const dayjs = require('dayjs');
-    import { getDateDiffList, formatTime } from '../js/common.js';
-    import FusionCharts from 'fusioncharts/core';
-    //import Bar2D from 'fusioncharts/viz/bar2d';
-    import StackedBar2D from 'fusioncharts/viz/stackedbar2d';
-    import Pie2D from 'fusioncharts/viz/pie2d';
-    import MSLine from 'fusioncharts/viz/msline';    
-    import InverseMSLine from 'fusioncharts/viz/inversemsline';    
-    import CandyTheme from "fusioncharts/themes/es/fusioncharts.theme.candy";
-    FusionCharts.addDep(StackedBar2D, Pie2D, MSLine, InverseMSLine, CandyTheme);
     
     export default {
         name: "SpeedRunGridChartsVue",
@@ -70,35 +82,33 @@
             return {
                 tabledata: [],
                 loading: true,
-                gameWorldRecordChart: {},
-                gameSpeedRunsPercentileChart: {},
-                gameTopSpeedRunChart: {}
+                showChartModal: false,
+                selectedChartID: 0
             }
-        },        
-        computed: {
-            isMediaMedium: function () {
-                return window.innerWidth > 768;
-            },
-            // isDarkTheme: function() {
-            //     return document.body.classList.contains('theme-dark');
-            // },            
-            bgColor: function () {
-                return document.body.classList.contains('theme-dark') ? "#303030" : "#f8f9fa";
-            },
-            fontColor: function () {
-                return document.body.classList.contains('theme-dark') ? "#fff" : "#212529";
-            }                           
-        },              
+        },
+        computed: {                 
+            chartModalTitle: function () {
+                var title = '';
+
+                switch(this.selectedChartID){
+                    case 1:
+                        title = this.isgame ? 'World Records Chart' : 'Personal Bests Chart';
+                        break;
+                    case 2:
+                        title = 'Time Percentiles Chart';
+                        break;
+                    case 3:
+                        title = 'Top 10 Chart';
+                        break;
+                }
+
+                return title;
+            }                                   
+        },                                     
         mounted: function () {
             this.loadData();
-        },
-        methods: {
-            // bgColor() {
-            //     return document.body.classList.contains('theme-dark') ? "#303030" : "#f8f9fa";
-            // },  
-            // fontColor() {
-            //     return document.body.classList.contains('theme-dark') ? "#fff" : "#212529";
-            // },                         
+        },      
+        methods: {                        
             loadData() {
                 var that = this;
                 this.loading = true;
@@ -106,420 +116,16 @@
                 axios.get('/SpeedRun/GetSpeedRunGridData', { params: { gameID: this.gameid, categoryID: this.categoryid, levelID: this.levelid, subCategoryVariableValueIDs: this.variablevalues, userID: this.userid, showAllData: true } })
                     .then(res => {
                         that.tabledata = res.data;                                             
-                        that.loadCharts();  
                         that.loading = false;  
                     })
                     .catch(err => { console.error(err); return Promise.reject(err); });
-            },            
-            loadCharts() {
-                var that = this;
-                FusionCharts.ready(function () {
-                    that.gameWorldRecordChart = new FusionCharts(that.getWorldRecordPerDayChart('divChart1', that.isgame));
-                    that.gameSpeedRunsPercentileChart = new FusionCharts(that.getSpeedRunsPercentileChart('divChart2'));
-                    that.gameTopSpeedRunChart = new FusionCharts(that.getTopSpeedRunChart('divChart3', that.isgame));
-
-                    that.gameWorldRecordChart.render();
-                    that.gameSpeedRunsPercentileChart.render();
-                    that.gameTopSpeedRunChart.render();                        
-                });
-            },
-            changeTheme() {
-                this.gameWorldRecordChart.bgColor = this.bgColor;                
-                this.gameWorldRecordChart.baseFontColor = this.fontColor;
-                this.gameWorldRecordChart.baseFontColor = this.fontColor;      
-                
-                this.gameSpeedRunsPercentileChart.bgColor = this.bgColor;                
-                this.gameSpeedRunsPercentileChart.baseFontColor = this.fontColor;
-                this.gameSpeedRunsPercentileChart.baseFontColor = this.fontColor;       
-      
-                this.gameTopSpeedRunChart.bgColor = this.bgColor;               
-                this.gameTopSpeedRunChart.baseFontColor = this.fontColor;
-                this.gameTopSpeedRunChart.baseFontColor = this.fontColor;    
-                
-                // this.gameWorldRecordChart.render();
-                // this.gameSpeedRunsPercentileChart.render();
-                // this.gameTopSpeedRunChart.render();               
-            },
-            getWorldRecordPerDayChart(container, isGame) {
-                var that = this;
-                var categories = [];
-                var dataset = [];
-
-                if (this.tabledata?.length > 0) {
-                    var _data = JSON.parse(JSON.stringify(this.tabledata)); 
-                    if (this.istimerasc) {
-                        _data = _data.sort((a, b) => { 
-                            return b?.primaryTimeMilliseconds - a?.primaryTimeMilliseconds
-                        });
-                    } else {
-                        _data = _data.sort((a, b) => { 
-                            return a?.primaryTimeMilliseconds - b?.primaryTimeMilliseconds
-                        });
-                    }
-                                        
-                    var filteredData = [];
-                    while (_data.length > 0) {
-                        var item = _data[0];
-                        filteredData.push(item);
-
-                        if (this.istimerasc) {
-                            _data = _data.filter(x => x.dateSubmitted < item.dateSubmitted)
-                                        .sort((a, b) => { return b?.primaryTimeMilliseconds - a?.primaryTimeMilliseconds });
-                        } else {
-                            _data = _data.filter(x => x.dateSubmitted < item.dateSubmitted)
-                                        .sort((a, b) => { return a?.primaryTimeMilliseconds - b?.primaryTimeMilliseconds });
-                        }                       
-                    }
-
-                    if (isGame) {
-                        filteredData = filteredData.slice(0, 20);
-                    }
-
-                    var dates = filteredData.map(item => { return new Date(item.dateSubmitted) });
-                    var maxDate = dayjs(Math.max.apply(null, dates)).startOf('day').toDate();//.add(1, "days").toDate();
-                    var minDate = dayjs(Math.min.apply(null, dates)).startOf('day').toDate();
-                    //var minDate = dayjs(maxDate).add(-24, "months");
-                    //filteredData = filteredData.filter(x => { return new Date(x.dateSubmitted) >= minDate; });
-
-                    var _timePeriods = getDateDiffList("day", minDate, maxDate).map(x => { return dayjs(x).format("MM/DD/YYYY") });
-
-                    var groupedObj = {};
-                    var chartDataObj = {};
-                    var categoryObj = {};
-
-                    filteredData.forEach(item => {
-                        var monthDayYear = dayjs(item.dateSubmitted).format("MM/DD/YYYY")
-                        var playerNames = item.players?.map(user => user.name).join("{br}");
-
-                        groupedObj[monthDayYear] = groupedObj[monthDayYear] || [];
-                        groupedObj[monthDayYear].push({ primaryTimeMilliseconds: item.primaryTimeMilliseconds, primaryTimeString: item.primaryTimeString, playerNames: playerNames });                        
-                    });
-
-                    if (Object.keys(groupedObj).length > 0) {
-                        for (var key in groupedObj) {
-                            chartDataObj[key] = chartDataObj[key] || [];
-
-                            var minItem = {};
-                            if (this.istimerasc) {
-                                minItem = groupedObj[key].sort((a, b) => { return b?.primaryTimeMilliseconds - a?.primaryTimeMilliseconds })[0];
-                            } else {
-                                minItem = groupedObj[key].sort((a, b) => { return a?.primaryTimeMilliseconds - b?.primaryTimeMilliseconds })[0];
-                            }
-
-                            chartDataObj[key] = { value: minItem.primaryTimeMilliseconds, tooltext: key + "{br}" + minItem.playerNames + "{br}" + minItem.primaryTimeString };                                                  
-                        }
-
-                        _timePeriods.forEach(timePeriod => {
-                            if (!chartDataObj.hasOwnProperty(timePeriod)) {
-                                chartDataObj[timePeriod] = { value: null, tooltext: ' ' };
-                            }
-                        });
-                    }
-
-                    categoryObj["category"] = _timePeriods.map(item => {
-                            var labelObj = {};
-                            labelObj["label"] = item;
-                            return labelObj;
-                        });
-                    categories.push(categoryObj);                    
-                        
-                    var chartData = Object.entries(chartDataObj)?.map(x => ({ label: x[0], value: x[1]?.value, tooltext: x[1]?.tooltext }));                        
-                    
-                    if (chartData.length > 0) {
-                        chartData = chartData.sort((a, b) => {
-                            var monthdayyeara = a.label.split("/");
-                            var monthdayyearb = b.label.split("/");
-
-                            return new Date(monthdayyeara[2], monthdayyeara[0] - 1, monthdayyeara[1]) - new Date(monthdayyearb[2], monthdayyearb[0] - 1, monthdayyearb[1])
-                        });
-                    }
-
-                    dataset.push({ seriesname: '', data: chartData });
+            },  
+            onChartClick(event, chartID) {
+                if (!event.target.innerHTML || event.target.innerHTML.indexOf("Export") == -1){
+                    this.showChartModal = true;
+                    this.selectedChartID = chartID;
                 }
-
-                const chartConfig = {
-                    type: this.istimerasc ? "inversemsline" : "msline",
-                    renderAt: container,
-                    width: "100%",
-                    dataFormat: "json",
-                    dataSource: {
-                        chart: {
-                            caption: isGame ? 'World Records (Last 20)' : 'Personal Bests',
-                            captionFontSize: 14,                           
-                            subCaption: that.title,
-                            subCaptionFontSize: 11,
-                            xAxis: 'Date',
-                            yAxis: 'Time (Minutes)',
-                            canvasPadding: 5, 
-                            labelDisplay: "ROTATE",
-                            labelFontSize: 11,
-                            showLabels: 1,
-                            rotateLabels: 1,
-                            slantLabels: 1,
-                            showToolTip: 1,
-                            showLegend: 0,
-                            lineThickness: 2,
-                            anchorRadius: 5,
-                            anchorBgColor: that.bgColor,
-                            anchorBorderThickness: 1,                                                   
-                            exportEnabled: 1,
-                            showValues: 0,
-                            formatNumberScale: 1,
-                            numberOfDecimals: 0,
-                            numberscalevalue: "1000,60,60",
-                            numberscaleunit: "s,m,h",
-                            defaultnumberscale: "ms",
-                            scalerecursively: "1",
-                            maxscalerecursion: "-1",
-                            scaleseparator: " ",
-                            connectNullData: 1,
-                            plotBinSize: 1.5,
-                            setAdaptiveYMin: 1,
-                            theme: "candy",
-                            bgColor: that.bgColor,
-                            baseFontColor: that.fontColor,
-                            outCnvBaseFontColor: that.fontColor
-                        },
-                        categories: categories,
-                        dataset: dataset
-                    }
-                };
-
-                return chartConfig;
-            },
-            getSpeedRunsPercentileChart(container) {
-                var that = this;
-                var dataset = [];
-
-                if (this.tabledata?.length > 0) {
-                    var _data = JSON.parse(JSON.stringify(this.tabledata)); 
-                    var allSpeedRunTimes = [];
-                    if (this.istimerasc) {
-                        allSpeedRunTimes = _data.sort((a, b) => { return b?.primaryTimeMilliseconds - a?.primaryTimeMilliseconds; });
-                    } else {
-                        allSpeedRunTimes = _data.sort((a, b) => { return a?.primaryTimeMilliseconds - b?.primaryTimeMilliseconds; });
-                    }
-
-                    var chartDataObj = {};
-                    var percIncrement = 5;
-                    var maxPerc = 25;
-                    var showEvery = 2;
-                    var maxNumCategories = Math.round((100 / percIncrement) / showEvery) + 1;
-
-                    var prevPercNum = null;
-                    var prevIndex = null;
-                    var prevTime = null;
-
-                    var prevTotal = 0;
-                    for (var i = 0; i < maxNumCategories; i++) {
-                        var percNum = (i == 0) ? percIncrement : prevPercNum + (percIncrement * showEvery);
-                        var index = Math.ceil((allSpeedRunTimes.length + 1) * (percNum / 100));
-                        index = ((index > 0) ? index - 1 : 0);// + ((prevIndex > 0) ? prevIndex - 1 : 0)
-
-                        var time;
-                        var key;
-                        var percent;
-                        var values = allSpeedRunTimes.filter((x, i) => { return i <= index });
-
-                        if (index >= allSpeedRunTimes.length - 1 || percNum > maxPerc || i == (maxNumCategories - 1)) {
-                            values = allSpeedRunTimes.length == 1 ? allSpeedRunTimes : allSpeedRunTimes.filter((x, i) => { return i > prevTotal });
-                            percent = Math.trunc((values.length / allSpeedRunTimes.length) * 100) || 0;
-                            key = (this.istimerasc ? '<= ' : '>= ') + formatTime("millisecond", prevTime) + " (" + percent + "% - " + values.length + "/" + allSpeedRunTimes.length + ")";
-                            chartDataObj[key] = values;
-                            break;
-                        } else {
-                            time = allSpeedRunTimes[index].primaryTimeMilliseconds;
-                            percent = Math.trunc((values.length / allSpeedRunTimes.length) * 100) || 0;
-                            key = (this.istimerasc ? '> ' : '< ') + formatTime("millisecond", time) + " (" + percent + "% - " + values.length + "/" + allSpeedRunTimes.length + ")";
-
-                            if (index != prevIndex) {
-                                chartDataObj[key] = values;
-                            }
-                        }
-
-                        prevTotal = values.length - 1;
-                        prevPercNum = percNum;
-                        prevIndex = index;
-                        prevTime = time;
-                    }
-
-                    dataset = Object.entries(chartDataObj)
-                        .map(x => {
-                            return { label: x[0], value: x[1].length }
-                        });
-                }
-
-                const chartConfig = {
-                    type: "pie2d",
-                    renderAt: container,
-                    width: "100%",
-                    dataFormat: "json",
-                    dataSource: {
-                        chart: {
-                            caption: 'Time Percentiles',
-                            captionFontSize: 14,                            
-                            subCaption: that.title,
-                            subCaptionFontSize: 11,
-                            showValues: 1,
-                            legendItemFontSize: 12,
-                            formatNumberScale: 0,
-                            numberOfDecimals: 0,
-                            showPercentValues: 0,
-                            showPercentInTooltip: 0,
-                            exportEnabled: 1,
-                            showLegend: 1,
-                            showLabels: 0,
-                            theme: "candy",
-                            bgColor: that.bgColor,
-                            baseFontColor: that.fontColor,
-                            outCnvBaseFontColor: that.fontColor
-                        },
-                        data: dataset
-                    }
-                };
-
-                return chartConfig;
-            },
-            // getTopSpeedRunChart(container, isGame) {
-            //     var dataset = [];
-
-            //     if (this.tabledata?.length > 0) {
-            //         var _data = JSON.parse(JSON.stringify(this.tabledata)); 
-
-            //         if (isGame) {
-            //             _data = _data.filter(x => x.rank);
-            //         }
-
-            //         var sortedData = _data.sort((a, b) => { return a?.primaryTimeMilliseconds - b?.primaryTimeMilliseconds; });
-            //         var data = sortedData.slice(0, 10);
-
-            //         dataset = data.map(item => {
-            //             return { label: item.players?.map(item => { return item.name; }).join("{br}"), 
-            //                      value: item.primaryTimeMilliseconds };
-            //         });
-            //     }
-
-            //     const chartConfig = {
-            //         type: "bar2d",
-            //         renderAt: container,
-            //         width: "100%",
-            //         dataFormat: "json",
-            //         dataSource: {
-            //             chart: {
-            //                 caption: isGame ? 'Top 10 Ranked' : 'Top 10',
-            //                 subCaption: '',
-            //                 xAxis: '',
-            //                 yAxis: 'Time (Minutes)',
-            //                 labelFontSize: 11,
-            //                 labelVAlign: 'middle',
-            //                 exportEnabled: 0,
-            //                 showValues: 1,
-            //                 placeValuesInside: 1,
-            //                 valueFontSize: 12,
-            //                 formatNumberScale: 1,
-            //                 numberOfDecimals: 0,
-            //                 useRoundEdges: 0,
-            //                 numberscalevalue: "1000,60,60",
-            //                 numberscaleunit: "s,m,h",
-            //                 defaultnumberscale: "ms",
-            //                 scalerecursively: "1",
-            //                 maxscalerecursion: "-1",
-            //                 scaleseparator: "",
-            //                 theme: "candy",
-            //                 bgColor: "#303030",
-            //                 baseFontColor: "#fff",
-            //                 outCnvBaseFontColor: "#fff"
-            //             },
-            //             data: dataset
-            //         }
-            //     };
-
-            //     return chartConfig;
-            // },
-            getTopSpeedRunChart(container, isGame) {
-                var that = this;                
-                var categories = [];
-                var dataset = [];
-
-                if (this.tabledata?.length > 0) {
-                    var _data = JSON.parse(JSON.stringify(this.tabledata)); 
-
-                    if (isGame) {
-                        _data = _data.filter(x => x.rank);
-                    }
-
-                    var sortedData = [];
-                    if (this.istimerasc){
-                        sortedData = _data.sort((a, b) => { return b?.primaryTimeMilliseconds - a?.primaryTimeMilliseconds; });
-                    } else {
-                        sortedData = _data.sort((a, b) => { return a?.primaryTimeMilliseconds - b?.primaryTimeMilliseconds; });
-                    }
-
-                    var data = sortedData.slice(0, 10);
-
-                    var chartDataObj = {};
-                    var categoryObj = {};
-                    data.forEach(item => {
-                        var playerNames = item.players?.map(user => user.name).join("{br}");
-
-                        chartDataObj[playerNames] = item.primaryTimeMilliseconds;
-                    });
-
-                    categoryObj["category"] = data.map(item => {
-                        var labelObj = {};
-                        labelObj["label"] = item.players?.map(item => {
-                            return item.name;
-                        }).join("{br}");
-                        return labelObj;
-                    });
-                    categories.push(categoryObj);
-
-                    var dataValues = data.map(item => {
-                        return { value: item.primaryTimeMilliseconds };
-                    });
-
-                    dataset.push({ seriesname: '', data: dataValues });
-                }
-
-                const chartConfig = {
-                    type: "stackedBar2D",
-                    renderAt: container,
-                    width: "100%",
-                    dataFormat: "json",
-                    dataSource: {
-                        chart: {
-                            caption: 'Top 10',
-                            captionFontSize: 14,                            
-                            subCaption: that.title,
-                            subCaptionFontSize: 11,
-                            xAxis: '',
-                            yAxis: 'Time (Minutes)',
-                            labelFontSize: 11,
-                            labelVAlign: 'middle',                            
-                            exportEnabled: 1,
-                            showValues: 1,
-                            valueFontSize: 11,
-                            formatNumberScale: 1,
-                            numberOfDecimals: 0,
-                            useRoundEdges: 0,
-                            numberscalevalue: "1000,60,60",
-                            numberscaleunit: "s,m,h",
-                            defaultnumberscale: "ms",
-                            scalerecursively: "1",
-                            maxscalerecursion: "-1",
-                            scaleseparator: "",
-                            theme: "candy",
-                            bgColor: that.bgColor,
-                            baseFontColor: that.fontColor,
-                            outCnvBaseFontColor: that.fontColor
-                        },
-                        categories: categories,
-                        dataset: dataset
-                    }
-                };
-
-                return chartConfig;
-            }            
+            }               
         }
     }
 </script>
