+++ SpeedRunApp.UI/src/components/component-loader.js
@@ -21,6 +21,9 @@ import speedRunGridTabVue from './SpeedRunGridTabVue.vue';
 import speedRunGridTabVariableVue from './SpeedRunGridTabVariableVue.vue';
 import speedRunGridVue from './SpeedRunGridVue.vue';
 import speedRunGridChartVue from './SpeedRunGridChartVue.vue';
+import worldRecordChartVue from './charts/WorldRecordChartVue.vue';
+import speedRunPercentileChartVue from './charts/SpeedRunPercentileChartVue.vue';
+import topSpeedRunChartVue from './charts/TopSpeedRunChartVue.vue';
 import worldRecordGridVue from './WorldRecordGridVue.vue';
 import worldRecordGridTabVue from './WorldRecordGridTabVue.vue';
 import userAccountVue from './UserAccountVue.vue';
@@ -53,6 +56,9 @@ export default {
         app.component("speedrun-grid-tab-variable", speedRunGridTabVariableVue);
         app.component("speedrun-grid", speedRunGridVue);
         app.component("speedrun-grid-chart", speedRunGridChartVue);
+        app.component("worldrecord-chart", worldRecordChartVue);
+        app.component("speedrun-percentile-chart", speedRunPercentileChartVue);
+        app.component("top-speedrun-chart", topSpeedRunChartVue);
         app.component("worldrecord-grid-tab", worldRecordGridTabVue);
         app.component("worldrecord-grid", worldRecordGridVue);
         app.component("personalbest-grid-tab", worldRecordGridTabVue);
