+++ SpeedRunApp.UI/src/styles/style.css
@@ -173,6 +173,14 @@ a.text-secondary:focus, a.text-secondary:hover {
     color: var(--bs-grey) !important;
 }
 
+.form-control::-moz-placeholder {
+    color: var(--bs-grey) !important;
+}
+
+.form-control:-ms-input-placeholder {
+    color: var(--bs-grey) !important;
+}
+
 .form-control::-ms-input-placeholder {
     color: var(--bs-grey) !important;
 }
@@ -183,7 +191,7 @@ a.text-secondary:focus, a.text-secondary:hover {
 
 .form-control:disabled, .form-control[readonly] {
     color: var(--bs-grey);
-    /* background-color: var(--bs-background-color2); */
+    -webkit-text-fill-color: var(--bs-grey);
     background-color: var(--bs-border-color10) !important;
 }
 
@@ -566,6 +574,10 @@ button.bg-dark:focus {
 }
 
 /** Fusioncharts Overrides **/
+.expandable .fusioncharts-container>svg:first-of-type {
+    cursor: pointer !important;
+}
+
 .fusioncharts-container>svg>g:nth-of-type(2) {
     display:none !important;
 }
@@ -1199,7 +1211,7 @@ input[readonly] {
 }
 
 .cmv-modal-header .cmv-close {
-    padding: 1rem 1rem;
+    padding: .75rem .75rem;
     margin: -1rem -1rem -1rem auto;
 }   
 
@@ -1405,7 +1417,7 @@ button.cmv-close {
 
 @media (max-width: 768px) {
     .btn {
-        font-size:11px;
+        font-size:14px;
         padding:4px 6px;
     }
